Merge iocp-ssl-593-2
Author: exarkun
Reviewer: glyph
Fixes: #593

Add a `ProtocolWrapper`-based implementation of TLS and use it
to extend the IO Completion Ports reactor with implementations
of `IReactorSSL` and `ITLSTransport`.

One of the existing SSL test suites is also changed to remove
one of its base classes (and therefore some of the tests it
was running); these tests had nothing to do with SSL, they were
just redundant TCP tests.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@26919 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb